check disk available space when write, and add flock on server PUT, but flock only for a process, so need more work on pthread_mutex_t

+++ server/srvpi.h
@@ -53,7 +53,7 @@ public:
 	Database * getPDB();
 	string getClipath();
 	string getFilename();
-	unsigned long getFilesize();
+	unsigned long long getFilesize();
 	~SrvPI();
